+++ Frends.Zip.ExtractArchive/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.1.0] - 2024-10-22
+### Fixed
+- Fixed issue with rename option writing the extracted files to wrong directory.
+### Added
+- Added DeletetZipFileAfterExtract option to enable the deletion of the source file.
+
 ## [1.0.3] - 2022-02-23
 ### Changed
 - Added documentation link to task description